Manager and Registry design
