+++ common/src/main/kotlin/com/lambda/core/Loader.kt
@@ -2,26 +2,25 @@ package com.lambda.core
 
 import com.lambda.Lambda
 import com.lambda.Lambda.LOG
-import com.lambda.command.CommandManager
+import com.lambda.command.CommandRegistry
+import com.lambda.friend.FriendRegistry
 import com.lambda.graphics.renderer.gui.font.LambdaFont
 import com.lambda.gui.impl.clickgui.GuiConfigurable
 import com.lambda.interaction.PlayerPacketManager
 import com.lambda.interaction.RotationManager
 import com.lambda.module.ModuleRegistry
-import com.lambda.plugin.PluginRegistry
 import com.lambda.util.Communication.ascii
 import kotlin.system.measureTimeMillis
 
 object Loader {
     private val loadables = listOf(
         ModuleRegistry,
-        CommandManager,
+        CommandRegistry,
         RotationManager,
         PlayerPacketManager,
         LambdaFont.Loader,
-        GuiConfigurable,
-        FriendManager,
-        PluginRegistry,
+        GuiConfigurable, // TODO: Why is this a loadable ?
+        FriendRegistry,
     )
 
     fun initialize() {
