+++ common/src/main/kotlin/com/lambda/module/modules/BaritoneTest.kt
@@ -1,9 +1,10 @@
-package com.lambda.module
+package com.lambda.module.modules
 
 import baritone.api.BaritoneAPI
 import baritone.api.pathing.goals.GoalXZ
 import com.lambda.event.events.TickEvent
 import com.lambda.event.listener.SafeListener.Companion.listener
+import com.lambda.module.Module
 
 object BaritoneTest : Module(
     name = "BaritoneTest",
